pyproject.toml
[project]
name = "parakeet-api-server"
version = "0.1.0"
description = "Minimal API server"
authors = [{name = "lb", email = "lb@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi >=0.104.0",
    "numpy>=2.0.2",
    "python-multipart>=0.0.20",
    "soundfile>=0.13.1",
    "tqdm>=4.67.2",
    "uvicorn [standard] >=0.24.0",
    "torch",
    "librosa>=0.11.0",
    "nemo-toolkit[asr]>=1.23.0",
    "sherpa-onnx>=1.12.23",
]

[project.scripts]
parakeet-api-server = "parakeet_api_server.server:main"


[build-system]
requires = ["uv", "hatchling"]
build-backend = "hatchling.build"

